Free the sensitive-stack memory in sensitive_clear_stack()
Otherwise this memory (16kb) is never freed.  This isn't a problem for
the main task since this stack has the same lifetime as the application,
but can cause problems when using ephmeral tasks eg. the camera task.

Also put the 16kb stack into spiram, since the dram is already
constrained, especaially for the camera task.