Check network type consistent at get_xpub and get_receive_address
Use a macro for these common checks, and use for all calls that use the
wallet keys.
Also check when unlocking jade, just as an early check that the wallet
is consistent with the network (otherwise it isn't detected until the
first use after pin entry, which is less convenient).