When signing always require prevout script to be passed explicitly
It is requried for legacy p2sh inputs as well as witness inputs, and
while it could be extracted from the input tx for p2pkh, it makes for
cleaner code and interface to simply always pass it explicitly.

Also slight tweak to liquid case to make code consistent, which also
fixes an issue with liquid when not all inputs are ours and we may not
have the relevant prevout script for those inputs - ie. only expect
prevout scripts for inputs we are actually signing.