Add an 'Initialize' button to the setup screen for uninitialised unit
This button navigates to the mnemonic entry/recovery screens, so the
wallet initialisation can occur prior to connecting to a companion app.
The user also selects the initial connection type, and BLE is enabled if
required.
When we then connect, the user just needs to set the PIN to persist the
wallet.
Also, remove the 'menu' from the bottom of the screen.