Generate and check generic multisig scripts and addresses
Provided the multisig is already registered with Jade:
a. support the generation and display of multisig receive addresses
b. support the automatic verification/validation of tx change addresses

In both cases all bip32 paths must be provided based on the xpubs in
the named multisig registration, and any paths considered 'unusual'
will cause a warning to be displayed on the screen (and confirmation
requested).