During OTA, display and check compressed firmware file hash
If the compressed fw file hash is passed, it is displayed on the OTA
confirm screen, and if after the entire fw file is uploaded the computed
hash does not match, the OTA fails.
If hash is not passed, no additional check is made.

At the moment passing the fw file hash (and hence running this check)
is optional, but it will be made mandatory in a future release.