Refactor gui activity event handling to use newer [un]register functions
Have the activities own the wait_event_data_t arguments as they need
to have the same lifetime as the activity, as the events can be
re-registered (eg. activity becomes current again).
Also, do not unregister the esp-event when the wait_event_data_t is
freed - it should happen when we switched to a new activity.
Ensure access to the current activity and its event registrations is
protected by a mutex.