tests: improve reliability of 'test_too_much_input()' test
Do not fail if bad bytes are reported over multiple messages - ie. do
not assume all bad bytes will be reported together in one message, so
long as the bytes reported as bad are as exepcted.
Format input as cbor message, so it is not rejected early as 'invalid
message'.