Add calls and tests for ssh and gpg integration
Interface and documentation for the calls to get an identity pubkey,
sign an identity challenge and generate an ecdh shared key for ssh and
gpg integration.  As per SLIP-0013 and SLIP-0017.
These calls take and return uncompressed pubkeys, valid for the given
curve.

Initially only supports nist256p1 (secp256r1).