liquid: Do not insist on deterministic blinders when signing liquid txns
Remove the requirement that Jade only signs liquid txns where the
commitments were created using the deterministic blinding factors Jade
uses when it blinds, and ignore any hmac field in the commitments data.
When signing, just check that the passed unblinded values that the user
has checked and the passed blinding factors do indeed combine to make
the commitments we are signing.