tests: Update existing liquid tests and add newly generated testnet cases
Some of the older liquid test cases have inconsistent values for the
final commitment (value-commitment inconsistent with vbf).  These
pass as Jade does not verify this.  Ideally we would adjust the test vbf
to be appropriate for the commitment, but since that is not feasible
have instead adjusted the value-commitment to fit with the existing vbf.
NOTE: some of the more recent test cases are already consistent/valid.
Added some new test cases, freshly generated using green-cli and
testnet-liquid.