ota: Correct error message id when ota-delta errors (or is declined)
If during fw patching we get an error when copying unchanged parts of
the base firmware to the new partition and applying patch changes, do
not always send the error reply immediately, as we may have already
sent an 'ok' reply for the last diff data message.
If so, cache the error on the context and send it in reply to the next
ota-data or ota-complete message.