gui: allow gui activities to be created outside of the managed list
As it stands gui activities are always held in a list which can be
periodically cleaned up.  This change allows for the creation of
activities outside of this list - eg. long-lived activities like the
main 'Ready' screen, and potentially any particularly large activities
which may want freeing asap.
It also allows for the list of outstanding activities to be freed more
frequently (since it wouldn't contain long-lived activities).