qrmode: scan and verify a receive address from previously exported xpub
If receive addresses are generated externally after the xpub has been
exported from Jade via QR, those addresses can be scanned and verified
as belonging to this Jade.
Addresses are verified by brute-forcing the index element of the
exported xpub until the same scriptpubkey is generated.
Non-change paths are assumed.