wallet: add function to search for a multisig address
Change the parameters to the function that builds a singlesig script so
we pass the derived key in (rather than the bip32 path) - and derive the
child key externally.
Refactor that function that searches for a singlesig script to reuse
that function, and and add a similar search wrapper for multisig scripts.