qrmode ui: use 'tick' rather than '<-' on display-qr screens
'<-' implies 'back', whereas a tick mark is a bit more ambiguous, and
works for any sort of 'done with this screen' action.
eg. in pinserver logon it's more of a 'next' than a 'back' action - and
a 'tick' works well in both situations.
Also make the 'tick' button the default/initially-selected button.