multisig: make 'our signer' xpub check slightly less strict
When registering a multisig record, Jade verifies that at least one of
the cosigners is this wallet/signer, and that the xpub provided matches
that calculated for this signer using the derivation path given.
It appears some third-party wallet apps pass xpubs that are not
strictly correct - some of the ancialliary/metadata fields (eg. depth,
parent-fingerprint) can be incorrect.
This change relaxes the xpub verification to parse the xpub provided
and check only the pubkey and chain-code parts (rather than generating
a base58 xpub locally and using string-compare).