qrmode: do not consider 'all connections lost' when QRs are in use
Adjust changes made in f6220e21dcbd8448af185f655f4f38ea4ac5538a to
detect when 'all connections' have been lost.  We need a slight tweak
exclude QRs from this test, as the QR 'source' is not 'connected' at all
and so can't be tested in that manner.
Instead we test if the last used source has disconnected, when that
source was usb/serial or ble.