mnemonic: add option to restrict final mnemonic word to valid options
As well as restricting to words on the wordlist starting with any typed
letters, add the option to restrict the final word to those that will
result in a valid bip39 mnemonic - ie. ones with a valid checksum.
Offer this option to the user when in 'Advanced' wallet recovery.
When selected, when the screen to select from a short list of words
appears, start on a random word in that list (rather than the
alphabetically first word).