liquid: make output commitment overrides optional when signing
When signing liquid tx, the unblinded output info, blinding factors and
final commitments are passed - they are checked for consistency and the
commitments are copied into the tx output and signed.
This change makes passing the final commitments optional, so they can be
omitted if the desired commitments are already present in the blinded tx
output.
The unblinded data and the blinding factors remain mandatory.