temporary stack: wrap provided 'esp_execute_shared_stack_function()'
esp-idf provides a mechanism to run a function on a temporary stack.
However the function interface must be 'void fn(void)' which means any
data and result have to be passed in static variables.

Provide a wrapper to present an interface 'bool fn(void* ctx)' which is
far easier to use.  Also protect the whole mechanism with a mutex so we
know only one 'large stack' can be allocated at any given time.