messaging: handle a non-blocking 'ping' message to check Jade powered
Handle a 'ping' message immediately in the messaging task to verify the
connection and that Jade is powered and receiving data, which returns
whether the main thread is currently busy handling a message, handling
user menu navigation, or is idle.

NOTE: this message is not queued for processing in fifo order, but is
handled immediately, even if the main task is busy.
This means if this message is sent while jade is processing other
messages, the replies may be out of order (although the 'id' field will
still be correct).  Use with caution.

NOTE: also added experimental version of 'get_version_info' which is
also handled immediately.