consistency: only return state as 'READY' if unlocked by this interface
Ensure state only shows 'READY' if this connection interface is
unlocked.  eg. if Jade is connected by serial but is unlocked and in-use
via BLE, the state via BLE is 'READY', but the state via serial is
'LOCKED'.
This is consistent with trying to sign over the 'other' interface,
which would return a 'LOCKED' error.