jadepy: make tcp reader await requested number of bytes
I believe this is the correct behaviour of the read() function.  At the
moment the wrapper returns with the number of bytes received so far,
even when that is less than that requested, thus relying on the caller
to loop and call again - which does not appear to be the standard
behaviour of the file-like read() function.