taproot: use SIGHASH_DEFAULT as the default sighash type for taproot inputs
Under taproot, SIGHASH_DEFAULT has the same signing semantics as SIGHASH_ALL,
but results in a 64 byte signature instead of 65, as the sighash byte is
omitted and inferred. Note that it is also possible to explicitly sign such
inputs with SIGHASH_ALL too.

Explicitly reject signing p2tr inputs for PSBTs with a FIXME until
implemented, now that we detect them.