sign_tx: move signing_data_t to wallet.h
For taproot, we are about to add more arguments to the calls that compute
a signature hash and sign a tx. As these calls already have many args,
move the struct that contains them to wallet.h to pass as a single argument
instead.

Signing imposes 2 limits: one on the path length, and one on the id field
length. Both constants for those limits exist outside of the wallet codes
access, and in another design layer. Avoid using the constants in
wallet.h and instead add a compile-time assertion that the values match
what the calling signing code expects.

Arguably some of the wallet.h functionality could be split into a tx.h,
but that isn't done here.