tests: migrate singlesig psbt tests to mnemonic + descriptor-based
Adds missing p2pkh and PSBTv2 test cases, and FIXMEs for test cases that
(IMO) should be added.

We cannot create p2tr testcases using an hdseed with core, since legacy
core wallets do not support taproot. We also do not have the mnemonic
that generates the hdseed currently used by the tests.

As a result, we need to migrate the singlesig tests to a new mnemonic,
and regenerate the test cases. Do this first for the psbt tests:
future commits will migrate the other tests.

This change means that test cases can be directly generated by and reconciled
with bitcoin-core. The mnemonic TEST_MNEMONIC_SINGLE_SIG results in the
following standard bip44-style account descriptors (not including the internal
change descriptor at /1/*):

- pkh([59d1f3b0/44'/1'/0']tpubDC2Q4xK4XH72GLPFZoHCRsGfPkBjCCo8wSH1wanCB5dMZB4pp3kEaMhmei7z9uigeeUhJD67oFGPpD8imaa1t5f1ZeNWDpCzGvGByzL9g51/0/*)#wnencdcp
- wpkh([59d1f3b0/84'/1'/0']tpubDC2Q4xK4XH72GrZ79KK6j3KAKkXiyVmknrJo6kH7XXUGQZqimwzmXX3RUmHwedXAgJXMpsy1usyh5FGq3Rx68qBAcU2ADyWzQqKxs2K7Vh9/0/*)#jxlpst3h
- sh(wpkh([59d1f3b0/49'/1'/0']tpubDC2Q4xK4XH72HZU6z1d5Lq2rSp5kwMvbtCaf7SCRHCNckPFB1jQLNKEQuLSFPyNNSEtJKXM85p9koZ4XLbkKXWJn4yWYpwYLx1MtrH1aTzM/0/*))#9phkc88q
- tr([59d1f3b0/86'/1'/0']tpubDC2Q4xK4XH72Gow34bNSZpx7uPcg1gfu6hPACSSieETYzpWgywMLmi2Yz9STA2Nrif3Yav7jvkzSj8q3nDKjjQrEfRYckUj5jsadYCdCw1C/0/*)#9wpuzyss

Multisig descriptors are created by taking the p2wpkh descriptor as one
key, with the other derived from TEST_MNEMONIC as a bip44-style, p2wpkh wallet.
An example of such as descriptor for a p2sh 2of2 (sorted) multisig is:

- sh(sortedmulti(2,[59d1f3b0/84'/1'/0']tpubDC2Q4xK4XH72GrZ79KK6j3KAKkXiyVmknrJo6kH7XXUGQZqimwzmXX3RUmHwedXAgJXMpsy1usyh5FGq3Rx68qBAcU2ADyWzQqKxs2K7Vh9/0/*,[e3ebcc79/84'/1'/0']tpubDC2Q4xK4XH72Gae41fkTzHPCugZXvDDZJwEDzVT5osM5ejQULHVtUhbjFqUBqTnMX5RVwtkstpBvjdznWG89y3Kt484Ta3qxTFPPWoyTh8s/0/*))#yzlv2wvy"

As above, the change descriptor for this case is the same except 0/* is
replaced by 1/*.