rpc: leave the output byte pointer null if cbor field is zero-sized
Prior to this change, passing a zero-sized (i.e. empty) cbor parameter
results in a returned size of zero, but a non-NULL pointer value which
points at data beyond the field.

Existing call sites are inconsistent with respect to pointer and/or
length checking.

Callers using rpc_get_string_ptr() sometimes set the return values directly
into struct members, which makes validating further use of the struct
fields difficult as we can then pass invalid non-NULL pointers to other parts
of the code.

Potentially the non-NULL pointer returned in this case could be used to
indicate that an empty field is present. But: (a) this would be an
error-prone interface for callers, (b) no existing callers do this, and
(c) we have rpc_has_field_data() to explicitly check if a field is present
which is clearer/better to use if required.

With this change we always return either a valid {pointer, length} pair,
or {NULL, 0}. Callers thus only need to check either the length or
pointer and do not risk passing invalid pointers elsewhere.