jadepy: do not use the legacy flow for non-AE signing
Prior to this change, requesting non-AE signing would use the legacy
flow which is deprecated. We now instead use the AE flow without
opting-in to AE signatures, and return the results in the format that
the legacy flow did (i.e. just the signatures with no signer commitments).

This change is API compatible(*), and means that jadepy users will no
longer use the legacy flow which will facilitate its removal in the future.

However, we still need to be able to test the legacy flow until its
removal. Add a 'use_legacy' parameter to enable this specifically for
test_jade.py to use.

(*) There is a small difference: if legacy signing was used, AE host
data could previously be passed and would be ignored - this is now an
error. If a caller is affected it can only mean either (a) they
intended to use AE signing but are not, or (b) they are not validating
the resulting AE signatures (because they are not receiving signer
commitments). Both need to be fixed by the caller, and this is
unavoidable if we wish to make Schnorr AE signing available later.