gui: remove 'must be active nodes' check from gui_set_active()
Having an activity with no active nodes is allowed, so not sure why we
have this check.  And in any case the external caller may set all nodes
to inactive, before then setting the active node.
Note also it is essentially an n-squared algorithm as every node checks
the other nodes.  Best removed.