sign_psbt: migrate key iteration to an interator interface
Taproot keypath are stored separately and contain slightly different
information than standard keypaths. Push the iteration of keys behind an
iterator interface in order to hide this from callers when taproot
support is added.

This initial change does not address the direct references to keypaths
in the calling code, in order to make the iteration logic easy to
review.