abort: update abort impl to work with amalgamated builds
Use __builtin_unreachable() to indicate to the compiler/optimizer that
execution can't reach a given spot. Supplement this with __noreturn__
for jade_abort() to indicate that it never returns.

Finally, expose a prototype for __wrap_abort and include the assertion
header in the abort implementation file.

This is functionally a no-op for non amalgamated builds, but is required
when amalgamating as -Wl,--wrap=<foo> does not work in the source file
that <foo> is defined in.