psbt: update checks for iterating taproot inputs/outputs
Ensure that even if we only have one key to sign with, that we
ignore inputs/outputs with an associated script path.

This prevents signing keypath spends on taproot inputs with scripts,
util script path spending is supported.

Also fixes iteration on an invalid iterator, leaving it as a no-op with
the iterator remaining invalid.