tests: fix tx output asset/value commitments in liquid txs
An upcoming change will force the asset and value in trusted_commitments
(if provided) to match with the values present in the transaction
outputs.

This change is intended to ensure that we do not mutate the
transaction/sign a differently blinded version of it that the caller did
not intend.

Fix test cases where the transaction and commitment data are inconsistent.