libjade: enable building the Jade firmware as a native library
Implements an in-process instance of the Jade firmware. This can be
thought of as an emulated or virtual Jade device that runs in the address
space of the application that links to it.

Also expose a daemon process which allows using libjade from another
process using the existing jadepy API.

This initial implementation is highly experimental and incomplete, and
should under no circumstances be used beyond development and testing.

Co-authored-by: Lawrence Nahum <lawrence@greenaddress.it>