usb: don't hold the callback mutex during the callback
Note that the callback is called via trigger_event() from a single thread
only and so holding the mutex while calling it is pointless. If the callback
triggers code which attempts to unset the callback (for example, during error
handling), then this will deadlock.

Instead, use the mutex only to protect reading and writing the callback and
context variables themselves.