ota: make user validation set the ota and confirmation status
This makes confirmation status easier to reason about as it is only set
in one place after ota_init().

Following this, it is clear that if the decompressed data callback is initially
called with less than the full custom header, then user validaton will
never be called. Document the code that catches this case and how it can
happen.